requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
amqp==5.2.0
anyio==4.2.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.0.1
billiard==4.2.0
celery==5.3.6
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.2
cssselect2==0.7.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
Django==4.2.3
django-allauth==0.59.0
django-paypal==2.0
djangorestframework==3.14.0
exceptiongroup==1.2.0
fabric==3.2.1
frozenlist==1.4.1
gunicorn==21.2.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httpx==0.26.0
idna==3.4
invoke==2.2.0
kombu==5.3.4
lxml==4.9.3
multidict==6.0.4
mysql-client==0.0.1
mysqlclient==2.2.0
oauthlib==3.2.2
oscrypto==1.3.0
packaging==23.1
paramiko==3.2.0
paypal==1.2.5
pdfkit==1.0.0
Pillow==10.0.0
prompt-toolkit==3.0.41
psycopg2-binary==2.9.9
pycparser==2.21
pyHanko==0.20.0
pyhanko-certvalidator==0.23.0
PyJWT==2.8.0
PyNaCl==1.5.0
pypdf==3.15.2
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
qrcode==7.4.2
rabbitmq==0.2.0
redis==5.0.1
reportlab==3.6.13
requests==2.31.0
requests-oauthlib==1.3.1
sib-api-v3-sdk==7.6.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.1
urllib3==2.0.4
vine==5.1.0
wcwidth==0.2.12
webencodings==0.5.1
wkhtmltopdf==0.2
wrapt==1.15.0
xhtml2pdf==0.2.11
yarl==1.9.4
